chore: decouple config and auth package
