+++ pkg/admin/auth/oauth2_test.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package admin
+package auth
 
 import (
 	"fmt"
